llm-inference-service/requirements.txt

fastapi>=0.111
uvicorn[standard]>=0.30
httpx>=0.27
pydantic>=2.7
PyYAML>=6.0
prometheus_client>=0.20
opentelemetry-sdk>=1.26.0
opentelemetry-exporter-otlp>=1.26.0
opentelemetry-instrumentation-fastapi>=0.47b0
loguru>=0.7.2
locust>=2.29.1
